+++ app/src/main/res/layout/fragment_discover_tab.xml
@@ -1,117 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              android:layout_width="match_parent"
-              android:layout_height="match_parent">
-
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" >
-
-        <com.amap.api.maps2d.MapView
-            android:id="@+id/map"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent" />
-
-        <LinearLayout
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="top|right"
-            android:background="@color/colorTransparent"
-            android:orientation="vertical"
-            android:padding="5dp" >
-
-            <LinearLayout
                 android:layout_width="match_parent"
-                android:layout_height="wrap_content" >
+                android:layout_height="match_parent">
 
-                <Button
-                    android:id="@+id/clearMap"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_weight="0.5"
-                    android:text="@string/clear_map" />
-
-                <Button
-                    android:id="@+id/resetMap"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_weight="0.5"
-                    android:text="@string/reset_map" />
-            </LinearLayout>
-        </LinearLayout>
-    </FrameLayout>
-
-    <RelativeLayout
-        android:id="@+id/id_marker_info"
-        android:visibility="gone"
-        android:layout_width="fill_parent"
-        android:layout_height="220dp"
-        android:layout_alignParentBottom="true"
-        android:background="#CC4e5a6b"
-        android:clickable="true" >
-
-        <ImageView
-            android:id="@+id/info_img"
-            android:layout_width="fill_parent"
-            android:layout_height="150dp"
-            android:layout_marginBottom="10dp"
-            android:layout_marginLeft="12dp"
-            android:layout_marginRight="12dp"
-            android:layout_marginTop="10dp"
-            android:alpha="1.0"
-            android:clickable="true"
-            android:scaleType="fitXY"/>
-
-        <RelativeLayout
-            android:layout_width="fill_parent"
-            android:layout_height="50dp"
-            android:layout_alignParentBottom="true">
-
-            <LinearLayout
-                android:layout_width="fill_parent"
-                android:layout_height="wrap_content"
-                android:layout_centerVertical="true"
-                android:layout_marginLeft="20dp"
-                android:orientation="vertical" >
-
-                <TextView
-                    android:id="@+id/info_name"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="老米家泡馍小炒"
-                    android:textColor="#FFF5EB" />
-
-                <TextView
-                    android:id="@+id/info_distance"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="距离200米"
-                    android:textColor="#FFF5EB" />
-            </LinearLayout>
-
-            <LinearLayout
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_alignParentRight="true"
-                android:layout_centerVertical="true"
-                android:layout_marginRight="20dp"
-                android:orientation="horizontal" >
-
-                <ImageView
-                    android:id="@+id/iv_zan"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:onClick="zan"/>
-
-                <TextView
-                    android:id="@+id/info_zan"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_gravity="center"
-                    android:text="652"
-                    android:textColor="#FFF5EB" />
-            </LinearLayout>
-        </RelativeLayout>
-    </RelativeLayout>
+    <android.support.v7.widget.RecyclerView
+        android:id="@+id/discover_recycler_view"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:background="@color/bg_gray">
+    </android.support.v7.widget.RecyclerView>
 
 </RelativeLayout>
